Let's get this started.
Daylio is awesome. I've been very happy with it for the three years I've
used it.

I discovered Day One while curiously looking to see if Daylio supported
exports to Apple Journals. (Doesn't work on the iPad; lol gtfo here with
that.)

Day One is awesome.

It doesn't have mood tracking, but it does have an Apple Watch app (that
works amazingly well), supports audio and video recordings, multiple
journals and more.

It's just too good. However, Daylio and Day One's data schemas are
different, and Day One doesn't make it obvious how to adapt other
formats to theirs. (It fails silently when you try to import.)

So I built this. I decided that using Go for this would be easier than
trying to mess with `jq` and Bash-isms.

Signed-off-by: Carlos Nunez <13461447+carlosonunez@users.noreply.github.com>