Yeah, so about those CSV files...
So diary entries in the Daylio CSV file strip newlines for some reason,
which is super unfortunate. What was _not_ unfortunate was discovering
that the full backups that Daylio writes to iCloud are not only easily
discoverable with a Mac, but are also just ZIP files with photos and a
JSON representation of Daylio's database.

Given that I would have needed to trawl through the backup, I figured
that it made more sense to just redesign this app to use the backup
anyway.

That's what this commit does.

Signed-off-by: Carlos Nunez <13461447+carlosonunez@users.noreply.github.com>